Changes to badges and login. See descr
Now badges on DB include a description.
New structure for badges on Dart.
Now all the teams (excluding their members) and all the badges of the loggedUser are retrieved during the login process.
To get the members of a team, use the getTeam function in mongodb_service.
New backend function to generate badges description, which is now required.