Rides with datetime. See descr
Now rides' date are a DateTime object.

Dart mongodb_service functions use MongoDB.formatDate on the ride date before sending it to the backend.

Furthermore, ride fromJSON constructor uses MongoDB.parseDate  to parse the date string received from the backend response.

Each ride object has a convenience displayDate method which returns a "nice-to-see" representation of the date.

run the test 'date testing' to see the effects of the different methods.

Code is commented as a reference